Move RecordingData related classed to internal-api module
