+++ dd-java-agent/agent-profiling/profiling-auxiliary-ddprof/src/main/java/com/datadog/profiling/auxiliary/ddprof/AuxiliaryDatadogProfiler.java
@@ -2,10 +2,10 @@ package com.datadog.profiling.auxiliary.ddprof;
 
 import com.datadog.profiling.auxiliary.AuxiliaryImplementation;
 import com.datadog.profiling.controller.OngoingRecording;
-import com.datadog.profiling.controller.RecordingData;
 import com.datadog.profiling.ddprof.DatadogProfiler;
 import com.datadog.profiling.utils.ProfilingMode;
 import com.google.auto.service.AutoService;
+import datadog.trace.api.profiling.RecordingData;
 import datadog.trace.bootstrap.config.provider.ConfigProvider;
 import java.util.EnumSet;
 import java.util.Set;
