+++ dd-java-agent/agent-profiling/profiling-auxiliary/src/main/java/com/datadog/profiling/auxiliary/AuxiliaryRecordingData.java
@@ -1,7 +1,7 @@
 package com.datadog.profiling.auxiliary;
 
-import com.datadog.profiling.controller.RecordingData;
-import com.datadog.profiling.controller.RecordingInputStream;
+import datadog.trace.api.profiling.RecordingData;
+import datadog.trace.api.profiling.RecordingInputStream;
 import java.io.IOException;
 import java.time.Instant;
 import javax.annotation.Nonnull;
