+++ dd-java-agent/agent-profiling/profiling-controller-ddprof/src/test/java/com/datadog/profiling/controller/ddprof/DatadogProfilerControllerTest.java
@@ -1,9 +1,9 @@
 package com.datadog.profiling.controller.ddprof;
 
 import static datadog.trace.api.config.ProfilingConfig.PROFILING_AUXILIARY_TYPE;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.datadog.profiling.controller.RecordingData;
+import datadog.trace.api.profiling.RecordingData;
 import datadog.trace.bootstrap.config.provider.ConfigProvider;
 import java.util.Properties;
 import org.junit.jupiter.api.Test;
