+++ dd-java-agent/agent-profiling/profiling-controller-openjdk/src/main/java/com/datadog/profiling/controller/openjdk/OpenJdkOngoingRecording.java
@@ -3,10 +3,10 @@ package com.datadog.profiling.controller.openjdk;
 import com.datadog.profiling.auxiliary.AuxiliaryProfiler;
 import com.datadog.profiling.auxiliary.AuxiliaryRecordingData;
 import com.datadog.profiling.controller.OngoingRecording;
-import com.datadog.profiling.controller.RecordingData;
 import com.datadog.profiling.utils.ProfilingMode;
 import datadog.trace.api.profiling.ProfilingListenersRegistry;
 import datadog.trace.api.profiling.ProfilingSnapshot;
+import datadog.trace.api.profiling.RecordingData;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Map;
