+++ dd-java-agent/agent-profiling/profiling-controller-openjdk/src/test/java/com/datadog/profiling/controller/openjdk/OpenJdkOngoingRecordingTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.datadog.profiling.controller.RecordingData;
+import datadog.trace.api.profiling.RecordingData;
 import java.time.Instant;
 import jdk.jfr.Recording;
 import jdk.jfr.RecordingState;
