+++ internal-api/src/main/java/datadog/trace/api/profiling/RecordingData.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.datadog.profiling.controller;
+package datadog.trace.api.profiling;
 
-import datadog.trace.api.profiling.ProfilingSnapshot;
 import java.io.IOException;
 import java.time.Instant;
 import javax.annotation.Nonnull;
